Add number extension
One may now extend a number by adding ".." to either end of a number, or
in the middle if the digits on both sides of the ".." are the same. The
number is extended up to the next-highest power of 2, for now. I fully
intend to make this more useful later, so it's not exposed as a public
function yet.